Create audienceType.ttl
This is a draft vocabulary for the audience types for use with LRMI properties. The skos:Concept instances in the vocabulary are drawn from existing audience vocabularies, have not been vetted by the LRMI community, and are subject to change. The purpose of the audience value vocabulary encoding is to provide an example of the value of a machine encoding using Simple Knowledge Organization System (SKOS) and its various concepts--not all of which are necessarily valid for this draft.