Update educationalAudienceRole.ttl
Added the <http://www.w3.org/2003/06/sw-vocab-status/ns#> namespace so we  can denote the status of terms at any particular stage. E.g., in this case, denoting all terms as "unstable" meaning: "The meaning, deployment practices, documentation (or important associated software/services) associated with this term are liable to change arbitrarily at some point in the future. They may not, but stability is not guaranteed. Use with caution."