Update educationalUse.ttl
Added <http://www.w3.org/2003/06/sw-vocab-status/ns#> namespace to support identifying the status of each term--here denoted as "unstable": "The meaning, deployment practices, documentation (or important associated software/services) associated with this term are liable to change arbitrarily at some point in the future. They may not, but stability is not guaranteed. Use with caution.
'testing'.